Renamed main python file; updated test file
